feat: :fire: add support for L510, L520, L610, L530 & L630 lights
In order to support L510, L520, L610, L530 & L630 lights, we had to modify the input format file to make it simpler on the user, then we modified the way we read this file and the associated code to initialize proper Tapo objects and communicate TouchPortal the right actions for each light. The previous read behaviour of the file remains 'the same' in a sense that if a light is present twice in a file, the newest like overrides the oldest one.

Breaks previous reading file format